Extract masterbar-visibility and section from ui.
The reducers and actions for state.ui.masterbar-visibility and
state.ui.section were part of the main state.ui modules. This change
splits them out, to allow for using them without depending on all of
state.ui.

Split-out actions are re-exported in the main file to maintain
compatibility for now.
