+++ client/state/ui/masterbar-visibility/actions.js
@@ -1,0 +1,18 @@
+/**
+ * Internal dependencies
+ */
+import { MASTERBAR_TOGGLE_VISIBILITY } from 'state/action-types';
+
+/**
+ * Hide the masterbar.
+ *
+ * @return {Object} Action object
+ */
+export const hideMasterbar = () => ( { type: MASTERBAR_TOGGLE_VISIBILITY, isVisible: false } );
+
+/**
+ * Show the masterbar.
+ *
+ * @return {Object} Action object
+ */
+export const showMasterbar = () => ( { type: MASTERBAR_TOGGLE_VISIBILITY, isVisible: true } );
