+++ client/state/ui/masterbar-visibility/reducer.js
@@ -1,0 +1,7 @@
+/**
+ * Internal dependencies
+ */
+import { MASTERBAR_TOGGLE_VISIBILITY } from 'state/action-types';
+
+export default ( state = true, { type, isVisible } ) =>
+	type === MASTERBAR_TOGGLE_VISIBILITY ? isVisible : state;
